Add files via upload
These should all be working, but you will have to edit the file locations for your own computer. As of now, the locations are for the RPi from the lab.

I was not sure about the format of the data, but right now it publishes to the /ordered_item rostopic as a string.